QemuVGADriver: add support for QEMU EDID data
Versions of QEMU >= 3.1 support passing of display information from host to guest
via an EDID blob accessible from the stdvga PCI MMIO BAR.

Whilst this feature is still in development, we can start by using the EDID data
to generate the list of supported guest resolutions as suggested by the host. As
EDID blob generation is currently disabled by default, it must be explicitly
enabled on the QEMU command line using "-device VGA,edid=on".

In the cases where the EDID data blob is not present, we fall back to using the
previous hard-coded set of guest resolutions.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>